# Changed the weight of the ruptures