The old weight was taking into consideration the number of realizations and assets per site, but since the performance is dominated by the hazard, the number of assets per site is irrelevant and taking it into account actually makes the tasks more heterogeneous. Mitigates the issue in https://github.com/gem/oq-engine/issues/2251.
Here are some numbers for the Chile computation in the cluster.

```
master
$ oq show performance 12545
=========================== ======== ========= =======
operation                   time_sec memory_mb counts 
=========================== ======== ========= =======
total event_based_risk      264,667  123       516    
building hazard             257,129  23        27,864 
total runtime               14,016   2,905     1      
computing risk              5,178    0.0       334,368

operation-duration mean  stddev min   max    num_tasks
event_based_risk   512   1,355  66    13,683 516      

this branch
$ oq show performance 12546
=========================== ======== ========= =======
operation                   time_sec memory_mb counts 
=========================== ======== ========= =======
total event_based_risk      264,157  106       516    
building hazard             256,544  15        27,864 
total runtime               10,290   2,896     1      
computing risk              5,309    0.0       334,368

operation-duration mean  stddev min   max   num_tasks
event_based_risk   511   925    63    9,991 516